requirements.txt
anycache==2.0.6
argon2-cffi==19.1.0
arrow==0.13.1
backcall==0.1.0
blinker==1.4
boto3==1.9.98
botocore==1.12.98
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
Click==7.0
cloudpickle==0.7.0
decorator==4.3.2
dill==0.2.9
docutils==0.14
filelock==3.0.10
Flask==1.0.2
Flask-Argon2==0.1.3.0
Flask-Mail==0.9.1
Flask-WTF==0.14.2
idna==2.8
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.10
jmespath==0.9.3
jupyter-client==5.2.4
jupyter-core==4.4.0
MarkupSafe==1.1.0
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.8
ptyprocess==0.6.0
py-s3-cache==1.0.0
pycparser==2.19
Pygments==2.3.1
python-dateutil==2.8.0
pyzmq==17.1.2
requests==2.21.0
s3transfer==0.2.0
six==1.12.0
sqlite3==0.0.0
Tkinter==0.0.0
tornado==5.1.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
Werkzeug==0.14.1
WTForms==2.2.1
